split test files up, remove responses.create stream logic
