+++ lib/instrumentation/openai.js
@@ -151,7 +151,9 @@ function recordChatCompletionMessages({
     // request.input can be a single string or an array of messages
     if (Array.isArray(request.input)) {
       messages = [...request.input, response?.output?.[0]]
-    } else messages = [request.input, response?.output?.[0]]
+    } else {
+      messages = [request.input, response?.output?.[0]]
+    }
   } else if (request.messages) {
     // `chat.completions.create` logic
     // Only take the first response message and append to input messages
@@ -228,16 +230,7 @@ function instrumentStream({ agent, shim, request, response, segment, transaction
         err = streamErr
         throw err
       } finally {
-        if (chunk.choices) {
-          // `chat.completions.create` logic
-          chunk.choices[0].message = { role, content }
-        } else if (chunk.response) {
-          // Note about `responses.create`:
-          // The final chunk has the response object
-          // in chunk.response, so we'll just use that
-          // as our response for recordChatCompletionMessages
-          chunk = chunk.response
-        }
+        chunk.choices[0].message = { role, content }
 
         // update segment duration since we want to extend the time it took to
         // handle the stream
@@ -296,7 +289,6 @@ module.exports = function initialize(agent, openai, moduleName, shim) {
 
   // Instruments chat completion creation
   // and creates the LLM events
-  // **Note**: Currently only for promises. streams will come later
   shim.record(
     openaiClient.Chat.Completions.prototype,
     'create',
@@ -339,6 +331,12 @@ module.exports = function initialize(agent, openai, moduleName, shim) {
   if (semver.gte(shim.pkgVersion, '5.0.0')) {
     shim.record(openaiClient.Responses.prototype, 'create', function wrapCreate(shim, create, name, args) {
       const [request] = args
+      if (request.stream) {
+        shim.logger.warn(
+          'Instrumenting responses.create streams is not yet supported.'
+        )
+        return
+      }
 
       return new RecorderSpec({
         name: OPENAI.COMPLETION,
