+++ test/versioned/openai/package.json
@@ -19,6 +19,17 @@
         "embeddings.test.js",
         "feedback-messages.test.js"
       ]
+    },
+    {
+      "engines": {
+        "node": ">=20"
+      },
+      "dependencies": {
+        "openai": ">=5.0.0"
+      },
+      "files": [
+        "chat-completions-v5.test.js"
+      ]
     }
   ]
 }
